+++ packages/http-helmet/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@mcansh/remix-secure-headers",
-  "version": "0.7.2",
+  "name": "@mcansh/http-helmet",
+  "version": "0.7.1",
   "description": "",
   "license": "MIT",
   "author": "Logan McAnsh <logan@mcan.sh> (https://mcan.sh)",
@@ -32,7 +32,8 @@
     "watch": "tsup --watch"
   },
   "dependencies": {
-    "@mcansh/http-helmet": "0.7.1"
+    "@remix-run/web-fetch": "^4.3.5",
+    "type-fest": "^3.13.0"
   },
   "publishConfig": {
     "access": "public"
