+++ packages/http-helmet/tsup.config.ts
@@ -1,12 +1,12 @@
 import { defineConfig } from "tsup";
 import type { Options } from "tsup";
-
 import pkgJson from "./package.json";
 
 let external = Object.keys(pkgJson.dependencies || {});
 
 export default defineConfig(() => {
   let options: Options = {
+    shims: true,
     entry: ["src/index.ts"],
     sourcemap: true,
     external,
